Added crab enemy to game

+++ PlatformGame/src/entities/Entity.java
@@ -1,8 +1,7 @@
- package entities;
+package entities;
 
 import java.awt.Color;
 import java.awt.Graphics;
-import java.awt.Rectangle;
 import java.awt.geom.Rectangle2D;
 
 public abstract class Entity {
@@ -26,7 +25,7 @@ public abstract class Entity {
 
 	}
 
-	protected void initHitbox(float x, float y, float width, float height) {
+	protected void initHitbox(float x, float y, int width, int height) {
 		hitbox = new Rectangle2D.Float(x, y, width, height);
 	}
